nlp-classification/pyproject.toml
[project]
name = "nlp-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jieba>=0.42.1",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "tensorboard>=2.20.0",
]
